First package version: 0.9.0
This packages everything into a Python module. Some files were combined (all Cobolt classes together, all COmmPort classes together, etc...)
