+++ hardwarelibrary/LaserSourceDevice.py
