+++ hardwarelibrary/__init__.py
@@ -1,0 +1,11 @@
+# We import everything by default, in the general namespace
+# because it is simpler for everyone
+
+from .CommunicationPort import *
+from .PhysicalDevice import *
+from .LaserSourceDevice import *
+from .Cobolt import *
+
+__version__ = "0.9.0"
+__author__ = "Daniel Cote <dccote@cervo.ulaval.ca>"
+
