+++ hardwarelibrary/findUSB.py
